Update environmental.py
There was a tiny but impactful error in this script. In line 443, I accidentally put 'Pressure':pres_variables[3][idx] instead of 'Pressure':pres_variables[2][idx] (as is now corrected). As a result, the coefficient-fit in line 506 went wrong. Anyone who has been trying to run STORM by themselves needs to check this in their own scripts and change accordingly.